ejemplo1/requirements.txt
transformers
langchain
langchain-huggingface
pillow
deep-translator